@@ -1,7 +1,7 @@
 /// for
 import '../../electrum_adapter.dart';
 
-extension GetAssetNamesMethod on RavenElectrumClient {
+extension GetAssetNamesMethod on FiroElectrumClient {
   Future<Iterable> getAssetsByPrefix(String symbol) async {
     return await request(
       'blockchain.asset.get_assets_with_prefix',
