@@ -1,5 +1,5 @@
 import '../../electrum_adapter.dart';
 
-extension PingServerMethod on RavenElectrumClient {
+extension PingServerMethod on FiroElectrumClient {
   Future<dynamic> ping() async => await request('server.ping');
 }
